+++ CHANGELOG.md
@@ -3,3 +3,15 @@
 ** 1.0.0 **
 
 * Initial Version
+
+** 1.1.0 **
+
+* Add orderCapture functionality
+* Add Webhook settings option
+* Add Test API-Key options in settings for 
+    * Klarna Pay Later
+    * Afterpay
+* Add Test IP Address options in settings for 
+    * Afterpay
+    * Klarna Pay Later
+* Add Country Validation